added database volume in docker compose

+++ deployment.sh
@@ -68,13 +68,14 @@ fi
 
 echo "Downloading configuration files..."
 
-sudo curl -L "https://raw.githubusercontent.com/WongSaang/chatgpt-ui/main/docker-compose.yml" -o docker-compose.yml
-
-echo "Pulling images..."
-
-sudo docker-compose pull
+# if docker-compose.yml doesn't exist, download it
+if [ ! -f "docker-compose.yml" ]; then
+    sudo curl -L "https://raw.githubusercontent.com/WongSaang/chatgpt-ui/main/docker-compose.yml" -o docker-compose.yml
+fi
 
 echo "Starting services..."
+
+touch ./db_sqlite3/db.sqlite3
 
 sudo APP_DOMAIN="${APP_DOMAIN}:${SERVER_PORT}" CLIENT_PORT=${CLIENT_PORT} SERVER_PORT=${SERVER_PORT} WSGI_PORT=${WSGI_PORT} DB_URL=${DATABASE_URL}  docker-compose up -d
